requirements.txt
ultralytics>=8.2.0
opencv-python
numpy
pandas
matplotlib
pyyaml
tqdm
scikit-learn
